% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awesome_cv.R
\name{awesome_cv}
\alias{awesome_cv}
\title{Create an Awesome-CV}
\usage{
awesome_cv(dir = file.path(getwd(), "my-awesome-cv"),
  color = "emerald", sections = c("education", "experience"))
}
\arguments{
\item{dir}{directory to place the new Awesome-CV}

\item{color}{awesome-cv color}
}
\description{
Create an Awesome-CV
}
