% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coursepackage.R
\name{cube}
\alias{cube}
\title{Cube a number}
\usage{
cube(x)
}
\arguments{
\item{x}{A numeric value to be cubed.}
}
\value{
A numeric value that is the cube of \code{x}.
}
\description{
Cube a number
}

