% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Interactive map of earthquakes}
\usage{
eq_map(.df, annot_col)
}
\arguments{
\item{.df}{(data.frame) earthquake data (NOAA)}

\item{annot_col}{(chr) name of the column used for the annotation.}
}
\value{
(leaflet) of earthquakes and annotations.
}
\description{
\code{eq_map} creates an interactive map
(\code{\link[leaflet]{leaflet}}) of earthquakes.
}
\examples{
\dontrun{
    library(dplyr)
    library(lubridate)
    library(devrcap)

    library(lubridate)
        data(noaa)

    noaa \%>\%
        eq_clean_data() \%>\%
        filter(
            country \%in\% c("ITALY", "GREECE", "PORTUGAL"),
            year(date) >= 1900
        ) \%>\%
        eq_map("location_name")


    noaa \%>\%
        eq_clean_data \%>\%
        dplyr::filter(country == "MEXICO", year(date) >= 2000) \%>\%
        eq_map("date")
}
}
