% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AbsPreGram.R
\name{find.nearest}
\alias{find.nearest}
\title{Find the nearest species}
\usage{
find.nearest(spec, gen)
}
\arguments{
\item{spec}{The number of the species in the set of the phylogenetic species}

\item{gen}{The number of the gene in the set of the genes read into the console}
}
\value{
The scientific name of the phylogenetically nearest species that has that gene available
}
\description{
Find the nearest species
}
\author{
Ali Amiryousefi
}

