% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Gw2Sw.R
\name{Gw2Sw}
\alias{Gw2Sw}
\title{Reformat the file from genwise to specieswise}
\usage{
Gw2Sw(dir = getwd(), duplicate = "random")
}
\arguments{
\item{dir}{The path where the fasta files are deposited}

\item{duplicate}{The method to handle the duplicate genes}
}
\description{
Function to convert the genewise fasta file of different species to specieswise file of different genes
}
\author{
Ali Amiryousefii
}

