% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_resume.R
\name{hugo_resume}
\alias{hugo_resume}
\title{Create a Hugo-Resume website}
\usage{
hugo_resume(dir = file.path(getwd(), "my-hugo-resume"),
  base_url = "https://www.mywebsite.com/")
}
\arguments{
\item{dir}{directory to place the new hugo-resume website}

\item{base_url}{base url for hosted website}
}
\description{
Create a Hugo-Resume website
}
