% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_resume.R
\name{hugo_resume_prepare_directory}
\alias{hugo_resume_prepare_directory}
\title{Prepare an new hugo-resume directory}
\usage{
hugo_resume_prepare_directory(dir)
}
\description{
Prepare an new hugo-resume directory
}
