% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nucl_nearest.R
\name{nucl_nearest}
\alias{nucl_nearest}
\title{Find the nearest species}
\usage{
nucl_nearest(spec)
}
\arguments{
\item{spec}{Is the species number}

\item{gen}{Is the gene of the speceis}
}
\description{
It will find and do the job
}
\author{
Ali Amiryousefi
}

