% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AbsPreGram.R
\name{plot.apg}
\alias{plot.apg}
\title{Plot the AbsPreGram}
\usage{
\method{plot}{apg}(tree, dir = getwd())
}
\arguments{
\item{tree}{the species tree}

\item{dir}{The directory where all the specieswise files of the genes are available. Note that all the files need to be in the fasta format with the '.fa' extention}
}
\description{
Plotting the matrix of the absense and the presence of the genes in the phylogeny
}
\author{
Ali Amiryousefi
}

