% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.R
\name{read_config}
\alias{read_config}
\title{Read the config.toml file containing info about me}
\usage{
read_config(file = system.file("me.toml", package = "me"))
}
\description{
Read the config.toml file containing info about me
}
