% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coursepackage.R
\name{square}
\alias{square}
\title{Square Function}
\usage{
square(x)
}
\arguments{
\item{x}{is a numeric value}
}
\value{
The numeric \code{value} that is the result of the square input
}
\description{
This the others.
}
\details{
This will squares the value of the numeric input
}
\examples{
x<-3
square(x) #returns the square value
}
\author{
Ali Amiryousefi
}
\seealso{
plot.apg
}

