% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitae.R
\name{vitae_cv}
\alias{vitae_cv}
\title{Create a CV using the `vitae` package}
\usage{
vitae_cv(..., template = "awesomecv", outdir = file.path(getwd(),
  "my_cv"), sections = c("education", "experience"))
}
\arguments{
\item{template}{which template should we use?}

\item{outdir}{full path to output directory}

\item{sections}{ordered vector of CV sections}
}
\description{
Create a CV using the `vitae` package
}
